feat: switch to google/mt5-base
This change introduces a switch to the more precise but also more complex mt5-base model.
With it, we introduce a repetition penalty of 1.25, a slight increase.
We also reduce the length penalty to 0 as the length is not that important.
The output/target length also increased to 128 tokens which allows for more questions to be generated.